[refactor] pipeline 구조 변경으로 인한 코드 수정
