+++ backend/src/main/java/com/example/backend/jenkins/job/repository/PipelineRepository.java
@@ -50,4 +50,6 @@ public interface PipelineRepository extends JpaRepository<Pipeline, UUID> {
     List<Pipeline> findDeletedWithScriptByJenkinsInfoId(@Param("jenkinsInfoId") UUID jenkinsInfoId);
 
     Optional<Pipeline> findByJenkinsInfoIdAndName(UUID jenkinsInfoId, String name);
+
+    Optional<Pipeline> findByJenkinsInfoIdAndId(UUID jenkinsInfoId, UUID jobId);
 }
